feat(ime): ✨ IME管理機能を追加
- IMEの状態を管理するための新しいモジュールを作成。
- グローバルIMEフックを設定し、IMEの状態を監視。
- IMEの状態変更を通知する機能を実装。
- システムトレイアイコンを追加し、ユーザーがアプリを終了できるようにした。
